added single food view
