+++ src/index.js
@@ -1,7 +1,7 @@
 import React from "react";
 import ReactDOM from "react-dom";
 
-import App from "./js/components/container/App.jsx";
+import App from "./js/components/App.jsx";
 import "bootstrap/dist/css/bootstrap.min.css";
 
 //if I want RenderDOM to be conditional
